refactor: colocate minigame logic with view components
