+++ src/components/views/minigames/star-trail/audio.js
