+++ src/components/views/minigames/star-trail/bgm.js
