+++ src/components/views/minigames/star-trail/core.js
