+++ src/components/views/minigames/star-trail/game.d.ts
