+++ src/components/views/minigames/star-trail/render.js
