+++ src/components/views/minigames/star-trail/style.css
