+++ src/components/views/minigames/star-trail/template.html
