+++ src/components/views/minigames/twilight/flandre.js
@@ -1,5 +1,5 @@
 // U.N. Owen piano reference by DMBN: original melody, inner harmony and bass registers.
-import score from "../../assets/minigames/twilight/un-owen-score.json";
+import score from "@/assets/minigames/twilight/un-owen-score.json";
 import { frequency } from "./harmony.js";
 
 const segments = [];
