+++ src/components/views/minigames/twilight/game.d.ts
